+++ board/sys_config/ti_msp_dl_config.c
@@ -150,8 +150,11 @@ SYSCONFIG_WEAK void SYSCFG_DL_GPIO_init(void)
 
     DL_GPIO_initPeripheralOutputFunction(
         GPIO_BlueTooth_IOMUX_TX, GPIO_BlueTooth_IOMUX_TX_FUNC);
-    DL_GPIO_initPeripheralInputFunction(
-        GPIO_BlueTooth_IOMUX_RX, GPIO_BlueTooth_IOMUX_RX_FUNC);
+    
+	DL_GPIO_initPeripheralInputFunctionFeatures(
+		 GPIO_BlueTooth_IOMUX_RX, GPIO_BlueTooth_IOMUX_RX_FUNC,
+		 DL_GPIO_INVERSION_DISABLE, DL_GPIO_RESISTOR_PULL_UP,
+		 DL_GPIO_HYSTERESIS_DISABLE, DL_GPIO_WAKEUP_DISABLE);
     DL_GPIO_initPeripheralOutputFunction(
         GPIO_Debug_Ex_IOMUX_TX, GPIO_Debug_Ex_IOMUX_TX_FUNC);
     DL_GPIO_initPeripheralInputFunction(